backend/requirements.txt
Flask==3.0.3
flask_cors==5.0.0
langchain==0.3.3
langchain_community==0.3.2
langchain_google_genai==2.0.1
open_clip_torch==2.23.0
Pillow==10.4.0
pymongo==4.10.1
python-dotenv==1.0.1
torch==2.4.1
faiss-cpu==1.9.0